move work to server
