+++ components/Search.js
@@ -1,66 +1,6 @@
 'use client';
-import React, { useState, useMemo } from 'react';
+import React, { useState } from 'react';
 import Link from 'next/link';
-
-/**
- * @typedef SearchTerms
- * @type {Map<string, Set<string>>}
- */
-
-function processTitle(title) {
-  return title.toLowerCase().replaceAll(' ', '-');
-}
-
-/**
- * @returns {SearchTerms} Map of words to episodes
- */
-function mapWordsToEpisodes(items) {
-  const searchTerms = new Map();
-  items.forEach(item => {
-    const { contentSnippet, title } = item;
-    const words = `${contentSnippet}${title}`
-      .toLowerCase()
-      .replace(/\n/g, ' ')
-      .replace(/[^\w ]*/g, '')
-      .split(' ');
-
-    words.forEach(word => {
-      if (searchTerms.has(word)) {
-        searchTerms.get(word).add(title);
-      } else {
-        searchTerms.set(word, new Set([title]));
-      }
-    });
-  });
-  return searchTerms;
-}
-
-/**
- * @param {string} wordStart
- * @param {SearchTerms} searchTerms
- */
-function getWords(wordStart, searchTerms) {
-  return [...searchTerms.keys()].filter(key => key.startsWith(wordStart));
-}
-
-/**
- * Takes a map of words to episodes and returns a trie
- * @param {SearchTerms} searchTerms
- */
-function createTrie(searchTerms) {
-  const trie = new Map();
-  for (const [word] of searchTerms) {
-    let node = trie;
-    word.split('').forEach((char, i) => {
-      if (!node.has(char)) {
-        node.set(char, new Map());
-      }
-      node = node.get(char);
-      node.set('words', getWords(word.substring(0, i + 1), searchTerms));
-    });
-  }
-  return trie;
-}
 
 function searchTrie(trie, searchTerm) {
   let node = trie;
@@ -75,12 +15,13 @@ function searchTrie(trie, searchTerm) {
   return node.get('words') || [];
 }
 
-function Search({ episodes }) {
-  const [searchTerm, setSearchTerm] = useState('');
+// TODO: Deduplicate
+function processTitle(title) {
+  return title.toLowerCase().replaceAll(' ', '-');
+}
 
-  // Only need to compute these once. TODO: Move this to the server
-  const searchTerms = useMemo(() => mapWordsToEpisodes(episodes), [episodes]);
-  const trie = useMemo(() => createTrie(searchTerms), [searchTerms]);
+function Search({ searchTerms, trie }) {
+  const [searchTerm, setSearchTerm] = useState('');
 
   const results = searchTrie(trie, searchTerm);
